Change ace to reconstruct
Replace ace with reconstruct. Ace sometimes has problems with reconstructing confidence intervals and returning NaN, which does not appear to be a problem with reconstruct. ace tentatively only used to calculate beta (which does not effect calculation of the D statistic at all) because reconstruct does not appear to currently return sigma2.